Modularized main.py into separate modules
