+++ .gitignore
@@ -4,3 +4,4 @@ polygon-test.py
 venv/*
 stock_analysis.log
 backup_main.py
+src/__pycache__/*
